DB migration P06 C4: REAL columns -> DOUBLE PRECISION (preserve float8 precision on Postgres) (#1126)
* P06 C4: declare REAL columns as DOUBLE PRECISION (preserve float8 precision on Postgres; SQLite unchanged)


* P06 C4: also fix par_level backfill CAST AS REAL -> DOUBLE PRECISION (PG branch)

---------

+++ dashboard/formulation_map.py
@@ -18,7 +18,7 @@ def init_tables(cx):
         min_age INTEGER, max_age INTEGER)""")
     cx.execute("""CREATE TABLE IF NOT EXISTS e4l_formulation_map(
         id INTEGER PRIMARY KEY AUTOINCREMENT, item_code TEXT, finding_pattern TEXT,
-        score_min REAL, formulation_id INTEGER, priority INTEGER,
+        score_min DOUBLE PRECISION, formulation_id INTEGER, priority INTEGER,
         other_therapies TEXT, notes TEXT, source TEXT)""")
     cx.commit()
